separate utility code
