+++ iOS/Example/MiniBeagle/Sources/Context.swift
@@ -22,62 +22,6 @@ struct Context {
     let value: Any
 }
 
-extension Collection {
-    subscript (safe index: Index) -> Element? {
-        return indices.contains(index) ? self[index] : nil
-    }
-}
-
-extension String {
-    func matches(pattern: String) -> [String] {
-        let regex = NSRegularExpression(pattern)
-        let results = regex.matches(in: self, range: NSRange(self.startIndex..., in: self))
-        return results.map {
-            return String(self[Range($0.range(at: 0), in: self)!])
-        }
-    }
-
-    func match(pattern: String) -> String? {
-        let regex = NSRegularExpression(pattern)
-        let result = regex.firstMatch(in: self, range: NSRange(self.startIndex..., in: self))
-        guard let unwrapped = result else { return nil }
-        return String(self[Range(unwrapped.range, in: self)!])
-    }
-}
-
-// MARK: BeagleExtensions
-// Observer pattern
-//protocol ObserverProtocol {
-//    var id : String { get }
-//}
-
-class Observable<T> {
-    typealias CompletionHandler = ((T) -> Void)
-    var value : T {
-        didSet {
-            self.notifyObservers()
-        }
-    }
-    var observers : [CompletionHandler] = []
-    init(_ value: T) {
-        self.value = value
-    }
-    func addObserver(completion: @escaping CompletionHandler) {
-        self.observers.append(completion)
-    }
-
-//    func removeObserver(_ observer: ObserverProtocol) {
-//        self.observers.removeValue(forKey: observer.id)
-//    }
-
-    func notifyObservers() {
-        observers.forEach { $0(value) }
-    }
-    deinit {
-        observers.removeAll()
-    }
-}
-
 //extension UIView: ObserverProtocol {
 //    var id: String {
 //        get {
@@ -97,11 +41,9 @@ extension UIView {
     var contextMap: [String: Observable<Context>]? {
         get {
             let contextMap: [String: Observable<Context>]? = (objc_getAssociatedObject(self, &UIView.contextMapKey) as? ObjectWrapper)?.object
-//            print("getContextMap: \(contextMap), object: \(self)")
             return contextMap
         }
         set {
-//            print("setContextMap: \(newValue), object: \(self)")
             objc_setAssociatedObject(self, &UIView.contextMapKey, ObjectWrapper(newValue), .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
         }
     }
@@ -109,14 +51,12 @@ extension UIView {
     func findContext(for expression: Expression) -> Observable<Context>? { // traversal
         // change to config binding
         guard let contextMap = self.contextMap else {
-//            print("parent: \(self.superview)")
             guard let parent = self.superview else {
                 return nil
             }
             return parent.findContext(for: expression)
         }
         guard let contextId = expression.context(), let context = contextMap[contextId] else {
-//            print("setContextMap: \(newValue), object: \(self)")
             guard let parent = self.superview else {
                 return nil
             }
